Search players
Two more themes are added (Black and White)
Search player implemented
Subfolders for screen

Player information is the main part of this program. With written code from before, it should not be that hard to implement with minor UI changes only.
For News part, a more button or something similar will be added. It will lead to useful tools & websites with two tabs to show users some recommended and official stuff. A browser will be written just to show a webview to render html for users with a open in broswer button top right.

Time to finish this app!